[#179] Rename `resolver` package to `ns`
`resolver` isn't suitable name for the package with NeoFS name system
functionality.
