+++ ns/doc.go
@@ -1,5 +1,5 @@
 /*
-Package resolver provides functionality of NeoFS name system.
+Package ns provides functionality of NeoFS name system.
 
 DNS type is designed to resolve NeoFS-related names using Domain Name System:
 	const containerName = "some-container"
@@ -19,4 +19,4 @@ NNS type is designed to resolve NeoFS-related names using Neo Name Service:
 	// ...
 
 */
-package resolver
+package ns
